types.ts move up from api
